Remove drive_control from dynamical decoupling module
